requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
jinja2==3.1.3
python-multipart==0.0.9
aiofiles==23.2.1
pydantic==2.6.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Text extraction
pypdf>=4.1.0
python-docx>=1.1.0

# LLM agent (provider-agnostic: OpenAI, Anthropic, Google, etc.)
litellm>=1.40.0
